feat: separation of classes into files. #172
